fix: moved PresenceType to presence/
