+++ src/presence/mod.rs
@@ -1,7 +1,36 @@
+use serde::{Deserialize, Serialize};
 use crate::{Client, RoboatError};
 
 const REGISTER_PRESENCE_API: &str = "https://presence.roblox.com/v1/presence/register-app-presence";
 
+/// Presence of user
+#[allow(missing_docs)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Serialize, Deserialize)]
+pub enum PresenceType {
+    #[default]
+    Offline,
+    Online,
+    InGame,
+    InStudio,
+    Invisible,
+}
+
+impl TryFrom<i32> for PresenceType {
+    type Error = RoboatError;
+
+    fn try_from(v: i32) -> Result<Self, Self::Error> {
+        match v {
+            0 => Ok(Self::Offline),
+            1 => Ok(Self::Online),
+            2 => Ok(Self::InGame),
+            3 => Ok(Self::InStudio),
+            4 => Ok(Self::Invisible),
+            _ => Err(RoboatError::MalformedResponse)
+        }
+    }
+}
+
+// TODO: add method for fetching users' presence
 impl Client {
     /// Registers presence on the website (makes you appear to be online). Endpoint called is
     /// <https://presence.roblox.com/v1/presence/register-app-presence>
